+++ firmware/src/worker.rs
@@ -342,6 +342,7 @@ impl<'a> Worker<'a> {
                 presence,
                 presence,
                 presence,
+                presence,
                 platform,
                 serial_id,
                 serial_hash,